Changed a few long to int

+++ LargestPrimeFactor/src/LargestPrimeFactor.java
@@ -13,14 +13,14 @@ public class LargestPrimeFactor {
 		
 	}
 	
-	public static boolean isPrime(long num){
+	public static boolean isPrime(int num){
 		if (num < 2)
 			return false;
 		if (num == 2)
 			return true;
 		if (num % 2 == 0)
 			return false;
-		for(long i = 3; i * i <= num; i += 2)
+		for(int i = 3; i * i <= num; i += 2)
 			if (num % i == 0)
 				return false;
 		return true;
